app/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
psycopg2-binary==2.9.10
psycopg2-pool==1.2
postgres==4.0

